requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.10
typing_extensions==4.5.0
Werkzeug==2.2.3
